FIX: Don't error out on non-stringy tags parameter (#180)
We're expecting the tags parameter to be a string. But the client can send an array or a nested parameter (thanks to the Rack protocol.) We're seeing exceptions in logs where the application is erroring out because of this.

As of this commit we raise a Discourse::InvalidParameters exception if we get a non-stringy tags parameter.