DEV: Use `transformedPost` as args to the Post widget (#202)
The post widgets expect to receive a transformed post as argument and use the spread operator to pass attributes to the inner widget, however the Docs plugin was providing a post model instance as argument instead.

This used to work fine, but as we introduce tracked properties in the post model, the spread operator doesn't copy their values, which causes issues.

This PR changes the `DocsTopic` Component to provide a `transformedPost` as argument to the post widget instead.

The change is necessary due to upcoming changes in the post model for the Glimmer post stream.